requirements.txt
flask
ultralytics
opencv-python-headless
numpy
pillow
gunicorn
Flask-Cors